Fix libappimage_static
